Move files around + line ending fixes
