+++ src/libsyntax/ext/expand.rs
@@ -14,9 +14,7 @@ use attr::{self, HasAttrs};
 use codemap::{ExpnInfo, NameAndSpan, MacroBang, MacroAttribute};
 use config::{is_test_or_bench, StripUnconfigured};
 use ext::base::*;
-use ext::build::AstBuilder;
-use ext::derive::{get_derive_attr, derive_attr_trait, add_derived_markers, verify_derive_attrs,
-                  DeriveType};
+use ext::derive::{find_derive_attr, derive_attr_trait};
 use ext::hygiene::Mark;
 use ext::placeholders::{placeholder, PlaceholderExpander};
 use feature_gate::{self, Features};
@@ -692,37 +690,9 @@ impl<'a, 'b> InvocationCollector<'a, 'b> {
         let mut attr = None;
 
         item = item.map_attrs(|mut attrs| {
-            attr = self.cx.resolver.find_attr_invoc(&mut attrs)
-                .or_else(|| {
-                    verify_derive_attrs(self.cx, &attrs);
-                    get_derive_attr(self.cx, &mut attrs, DeriveType::Legacy).and_then(|a| {
-                        let titem = derive_attr_trait(self.cx, &a);
-                        if let Some(titem) = titem {
-                            let tword = titem.word().unwrap();
-                            let tname = tword.name();
-                            if !self.cx.ecfg.enable_custom_derive() {
-                                feature_gate::emit_feature_err(&self.cx.parse_sess,
-                                                               "custom_derive",
-                                                               titem.span,
-                                                               feature_gate::GateIssue::Language,
-                                                               feature_gate::EXPLAIN_CUSTOM_DERIVE);
-                            } else {
-                                let name = Symbol::intern(&format!("derive_{}", tname));
-                                if !self.cx.resolver.is_whitelisted_legacy_custom_derive(name) {
-                                    self.cx.span_warn(titem.span, feature_gate::EXPLAIN_DEPR_CUSTOM_DERIVE);
-                                }
-                                let mitem = self.cx.meta_word(titem.span, name);
-                                return Some(self.cx.attribute(mitem.span, mitem));
-                            }
-                        }
-                        None
-                    })
-                }).or_else(|| {
-                    add_derived_markers(self.cx, &mut attrs);
-                    get_derive_attr(self.cx, &mut attrs, DeriveType::Builtin)
-                }).or_else(|| {
-                    get_derive_attr(self.cx, &mut attrs, DeriveType::ProcMacro)
-                });
+            attr = self.cx.resolver.find_attr_invoc(&mut attrs).or_else(|| {
+                find_derive_attr(self.cx, &mut attrs)
+            });
 
             attrs
         });
@@ -958,7 +928,8 @@ impl<'a, 'b> Folder for InvocationCollector<'a, 'b> {
 
         let (item, attr) = self.classify_item(item);
         if let Some(attr) = attr {
-            let item = Annotatable::TraitItem(P(fully_configure!(self, item, noop_fold_trait_item)));
+            let item =
+                Annotatable::TraitItem(P(fully_configure!(self, item, noop_fold_trait_item)));
             return self.collect_attr(attr, item, ExpansionKind::TraitItems).make_trait_items()
         }
